commit before editing structure of impurity_model

+++ continuous_impurity/toolbox/data_helper.py
@@ -1,6 +1,6 @@
 import numpy as np
 
 def affine_X(X):
-    X_affine = np.ones((X.shape[0], X.shape[1]+1))
+    X_affine = np.ones((X.shape[0], X.shape[1]+1), dtype = X.dtype)
     X_affine[:,:X.shape[1]] = X
     return X_affine
